Spring Secrurity implemented register, authenticate, refresh-token endpoints finished

+++ server/src/main/java/com/example/ideas/security/token/TokenRepository.java
@@ -8,14 +8,14 @@ import java.util.List;
 import java.util.Optional;
 
 @Repository
-public interface TokenRepository extends JpaRepository<Token, Integer> {
+public interface TokenRepository extends JpaRepository<Token, Long> {
 
   @Query(value = """
       select t from Token t inner join User u\s
       on t.user.userId = u.userId\s
       where u.userId = :id and (t.expired = false or t.revoked = false)\s
       """)
-  List<Token> findAllValidTokenByUser(Integer id);
+  List<Token> findAllValidTokenByUser(Long id);
 
   Optional<Token> findByToken(String token);
 }
